add guild subscription endpoints
