+++ docker/config-development.py
@@ -14,5 +14,8 @@ DISCORD_CLIENT_SECRET = os.getenv('DISCORD_CLIENT_SECRET')
 OAUTH_REDIRECT_URI = "http://127.0.0.1:4200/login"
 OAUTH_SCOPE = "identify guilds"
 
+# other discord
+DISCORD_BOT_TOKEN = os.getenv('DISCORD_BOT_TOKEN')
+
 # site auth
 JWT_SECRET = os.getenv('JWT_SECRET')
