+++ workshop/errors.py
@@ -9,3 +9,8 @@ class CollectionNotFound(AvraeException):
 class CollectableNotFound(AvraeException):
     def __init__(self, msg="collectible not found"):
         super().__init__(msg)
+
+
+class NeedsServerAliaser(AvraeException):
+    def __init__(self, msg="You do not have permissions to edit server collections on this server"):
+        super().__init__(msg)
